+++ packages/thea/src/sdk/evm/sepolia.ts
@@ -38,6 +38,10 @@ export class Sepolia implements EVMChainAdapter {
     });
   }
 
+  getAllAssets(): Asset[] {
+    return [ETHEREUM_ASSETS.ETH, ETHEREUM_ASSETS.LINK];
+  }
+
   getChain(): Chain {
     return sepoliaChain;
   }
@@ -46,7 +50,7 @@ export class Sepolia implements EVMChainAdapter {
     return [ETHEREUM_ASSETS.ETH, ETHEREUM_ASSETS.LINK];
   }
 
-  getDestinationChains(asset: Asset): Chain[] {
+  getDestinationChains(): Chain[] {
     const substrate_chains = [polkadex].map((c) =>
       changeSubstrateToBaseChain(c)
     );