+++ packages/ui/src/components/index.ts
@@ -40,3 +40,5 @@ export * from "./carousel";
 export * from "./interactable";
 export * from "./passcode";
 export * from "./progressBar";
+export * from "./chain";
+export * from "./intro";